# exec: add templating to vectorized merge join 