The second last step in the vectorized merge join operator is the
introduction of templating so that the equality type isn't limited
to just int64s.

With this change, there are also performance boosts as the
selection vector switching is now moved outside the tight loop.

Further, we noticed a bug in the previous completeGroups
implementation, so we killed two birds with one stone by both
templating and fixing that edge case with the use of the distinct
operator.

Special thanks to @jordanlewis who is really a coauthor of this
PR.

And as usual, some benchmarks:

```
name                                      old time/op    new time/op    delta
MergeJoiner/rows=1024-8                     44.9µs ± 1%    35.3µs ± 0%  -21.43%  (p=0.000 n=10+10)
MergeJoiner/rows=4096-8                      171µs ± 0%     133µs ± 0%  -22.15%  (p=0.000 n=10+10)
MergeJoiner/rows=16384-8                     656µs ± 1%     508µs ± 1%  -22.60%  (p=0.000 n=10+10)
MergeJoiner/rows=1048576-8                  40.3ms ± 1%    30.8ms ± 1%  -23.43%  (p=0.000 n=10+10)
MergeJoiner/oneSideRepeat-rows=1024-8       45.4µs ± 0%    35.4µs ± 0%  -22.19%  (p=0.000 n=10+9)
MergeJoiner/oneSideRepeat-rows=4096-8        171µs ± 0%     133µs ± 0%  -22.16%  (p=0.000 n=10+8)
MergeJoiner/oneSideRepeat-rows=16384-8       655µs ± 0%     508µs ± 0%  -22.54%  (p=0.000 n=10+10)
MergeJoiner/oneSideRepeat-rows=1048576-8    40.3ms ± 1%    31.0ms ± 1%  -23.12%  (p=0.000 n=10+9)
MergeJoiner/bothSidesRepeat-rows=1024-8     48.3µs ± 1%    37.6µs ± 1%  -22.25%  (p=0.000 n=10+10)
MergeJoiner/bothSidesRepeat-rows=4096-8      186µs ± 1%     174µs ± 0%   -6.05%  (p=0.000 n=10+10)
MergeJoiner/bothSidesRepeat-rows=16384-8    1.41ms ± 0%    1.39ms ± 1%   -1.59%  (p=0.000 n=10+10)
MergeJoiner/bothSidesRepeat-rows=32768-8    4.93ms ± 1%    5.28ms ± 1%   +7.17%  (p=0.000 n=9+10)

name                                      old speed      new speed      delta
MergeJoiner/rows=1024-8                   1.46GB/s ± 1%  1.86GB/s ± 0%  +27.28%  (p=0.000 n=10+10)
MergeJoiner/rows=4096-8                   1.53GB/s ± 0%  1.97GB/s ± 0%  +28.46%  (p=0.000 n=10+10)
MergeJoiner/rows=16384-8                  1.60GB/s ± 1%  2.06GB/s ± 1%  +29.20%  (p=0.000 n=10+10)
MergeJoiner/rows=1048576-8                1.67GB/s ± 1%  2.18GB/s ± 1%  +30.60%  (p=0.000 n=10+10)
MergeJoiner/oneSideRepeat-rows=1024-8     1.44GB/s ± 0%  1.85GB/s ± 0%  +28.52%  (p=0.000 n=10+9)
MergeJoiner/oneSideRepeat-rows=4096-8     1.54GB/s ± 0%  1.97GB/s ± 0%  +28.47%  (p=0.000 n=10+8)
MergeJoiner/oneSideRepeat-rows=16384-8    1.60GB/s ± 0%  2.07GB/s ± 0%  +29.10%  (p=0.000 n=10+10)
MergeJoiner/oneSideRepeat-rows=1048576-8  1.67GB/s ± 1%  2.17GB/s ± 1%  +30.07%  (p=0.000 n=10+9)
MergeJoiner/bothSidesRepeat-rows=1024-8   1.36GB/s ± 1%  1.74GB/s ± 1%  +28.62%  (p=0.000 n=10+10)
MergeJoiner/bothSidesRepeat-rows=4096-8   1.41GB/s ± 1%  1.50GB/s ± 0%   +6.43%  (p=0.000 n=10+10)
MergeJoiner/bothSidesRepeat-rows=16384-8   743MB/s ± 0%   755MB/s ± 1%   +1.62%  (p=0.000 n=10+10)
MergeJoiner/bothSidesRepeat-rows=32768-8   426MB/s ± 1%   397MB/s ± 1%   -6.69%  (p=0.000 n=9+10)

name                                      old alloc/op   new alloc/op   delta
MergeJoiner/rows=1024-8                     4.23kB ± 0%    4.23kB ± 0%   -0.07%  (p=0.000 n=10+10)
MergeJoiner/rows=4096-8                     4.25kB ± 0%    4.25kB ± 0%     ~     (all equal)
MergeJoiner/rows=16384-8                    4.34kB ± 0%    4.30kB ± 0%   -0.92%  (p=0.000 n=10+10)
MergeJoiner/rows=1048576-8                  12.2kB ± 0%     9.0kB ± 0%  -26.16%  (p=0.000 n=10+10)
MergeJoiner/oneSideRepeat-rows=1024-8       4.23kB ± 0%    4.23kB ± 0%   -0.07%  (p=0.000 n=10+10)
MergeJoiner/oneSideRepeat-rows=4096-8       4.25kB ± 0%    4.25kB ± 0%     ~     (all equal)
MergeJoiner/oneSideRepeat-rows=16384-8      4.34kB ± 0%    4.30kB ± 0%   -0.92%  (p=0.000 n=10+10)
MergeJoiner/oneSideRepeat-rows=1048576-8    12.2kB ± 0%     9.0kB ± 0%  -26.16%  (p=0.000 n=10+10)
MergeJoiner/bothSidesRepeat-rows=1024-8     4.23kB ± 0%    4.23kB ± 0%   -0.07%  (p=0.000 n=10+10)
MergeJoiner/bothSidesRepeat-rows=4096-8     4.25kB ± 0%    4.25kB ± 0%     ~     (all equal)
MergeJoiner/bothSidesRepeat-rows=16384-8    4.46kB ± 0%    4.46kB ± 0%     ~     (all equal)
MergeJoiner/bothSidesRepeat-rows=32768-8    5.02kB ± 0%    5.02kB ± 0%     ~     (all equal)

name                                      old allocs/op  new allocs/op  delta
MergeJoiner/rows=1024-8                       4.00 ± 0%      4.00 ± 0%     ~     (all equal)
MergeJoiner/rows=4096-8                       4.00 ± 0%      4.00 ± 0%     ~     (all equal)
MergeJoiner/rows=16384-8                      4.00 ± 0%      4.00 ± 0%     ~     (all equal)
MergeJoiner/rows=1048576-8                    6.00 ± 0%      5.00 ± 0%  -16.67%  (p=0.000 n=10+10)
MergeJoiner/oneSideRepeat-rows=1024-8         4.00 ± 0%      4.00 ± 0%     ~     (all equal)
MergeJoiner/oneSideRepeat-rows=4096-8         4.00 ± 0%      4.00 ± 0%     ~     (all equal)
MergeJoiner/oneSideRepeat-rows=16384-8        4.00 ± 0%      4.00 ± 0%     ~     (all equal)
MergeJoiner/oneSideRepeat-rows=1048576-8      6.00 ± 0%      5.00 ± 0%  -16.67%  (p=0.000 n=10+10)
MergeJoiner/bothSidesRepeat-rows=1024-8       4.00 ± 0%      4.00 ± 0%     ~     (all equal)
MergeJoiner/bothSidesRepeat-rows=4096-8       4.00 ± 0%      4.00 ± 0%     ~     (all equal)
MergeJoiner/bothSidesRepeat-rows=16384-8      4.00 ± 0%      4.00 ± 0%     ~     (all equal)
MergeJoiner/bothSidesRepeat-rows=32768-8      4.00 ± 0%      4.00 ± 0%     ~     (all equal)

```

Release note: None